Updated level 1, now has collision
Walls and player ball have collision, and a ball of light now sits in
the... arena? Anywho, the ball's movement is still not constant, so i'll
continue looking for a way to make it move at a constant speed in a
straight line in the direction clicked by the user.